services/tf-liquor-classifier/requirements.txt
tensorflow==2.3.*
numpy==1.18.*
matplotlib
boto3==1.17.*
pillow==7.2.*
cloudevents==1.2.*
mysql-connector-python==8.0.*
Flask==1.1.*
flask-cors==3.0.*